Add header to the table and clean up interest code

+++ src/LoanDetails.java
@@ -1,7 +1,7 @@
 import java.util.Date;
 
 public class LoanDetails {
-  Float interest;
+  Double interest;
   Double loanAmount;
   Integer term;
   Date requestDate;
